Script for inspecting/inserting links.
refs #5235

Created a CLI utility to make altering microservice chains easier.
It will, given a link UUID, show info about it (links coming before or
after it in the microservice group) and can ook up a link's UUID using
a task UUID. With a link's UUID you can also generate insertion SQL for
adding a new link before or after it in the microservice chain. The
script will prompt for link properties so they can be entered.