Update FPRClient to make more robust
refs #5825

Update FPR client to fix some bugs downloading rules, and add tests for
those cases.

When downloading a new rule (not replacing anything), always set it to
active.  If not, the entire chain is set to disabled.

When downloading a rule that has a local modification, insert the newly
downloaded rule immediately after the one it replaces in the modification
chain, updating the local modification to point at the newly downloaded
rule.  If not, the new rule does not appear in the GUI.

Rename main.py to client.py in the FPRClient because it conflicts with the
main folder in the dashboard when importing.  Update the installer for the
change.  Update fpradmin version.

Add tests for the FPRClient in test_fprclient.py.  Can be run with pytest.