Add PRONOM import script
This script takes a FIDO PRONOM XML file, and populates the FPR with
any new formats not yet present in the database. "New" formats are
defined as formats whose PUID is not present in any FormatVersion,
regardless of whether anything with a description of that name is
present.

If the FPR admin being used has debug enabled, then this will also
output the SQL statements used to insert entries into the database to
stdout. This makes it possible to generate a new SQL dump which can
be used to reimport those formats at a later point in time without
running this script - for instance, to upgrade an Archivematica's FPR
install without hitting the internet.

refs #6641